LayoutTests/fast/canvas/webgl/readPixels-float-expected.txt

Create WebGL context
Clear to black
Create a floating point texture
Create and bind framebuffer
Clear framebuffer to red
PASS: Framebuffer is floating-point red.
